Move controller to admin directory
